requirements.txt
asgiref==3.9.1
boto3==1.40.1
botocore==1.40.1
Django==5.2.5
django-environ==0.12.0
django-storages==1.14.6
dotenv==0.9.9
Faker==37.5.3
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
